ESTS-136762 Temporary fix for compute api change (#400)
As part of work to add the serial number, manufacturer and vendor to
the discovered nodes data the api was changed to include that
data. That changed the constructor arguments and broke DNE
compilation. This is a temporary change to adjust to those changes
until the full work is done.